Windows MSVC++ VC9 compability fix due to an "unrecoverable" error that occurs when fabs is invoked with integer value. Poor MSVC don't know which overloaded method to choose!
git-svn-id: https://bloopz.zapto.org/svn/nightfall/projekt/trunk@1011 d57568f3-fe0c-0410-b608-e86116ac1d00

+++ src/networking.cpp
@@ -312,7 +312,7 @@ namespace Game
 			while (curnode->pParent)
 			{
 				int stepcode = -1;
-				if (fabs(curnode->pParent->x - curnode->x) <= 1 && fabs(curnode->pParent->y - curnode->y) <= 1)
+				if (fabs((float)curnode->pParent->x - curnode->x) <= 1 && fabs((float)curnode->pParent->y - curnode->y) <= 1)
 				{
 					stepcode = stepcodes[(curnode->pParent->y - curnode->y)+1][(curnode->pParent->x - curnode->x)+1];
 				}
