+++ src/unit.cpp
@@ -419,7 +419,7 @@ namespace Game
 			}
 
 			int dist_x = attacker_x - x, dist_y = attacker_y - y;
-			int dist_both = fabs(dist_x) + fabs(dist_y);
+			int dist_both = fabs((float)dist_x) + fabs((float)dist_y);
 			if (dist_both <= maxrange)
 			{
 				if (dist_both < minrange)
@@ -428,16 +428,16 @@ namespace Game
 				}
 				else
 				{
-					distance = (float) sqrt(dist_x * dist_x + dist_y * dist_y);
+					distance = (float) sqrt(float(dist_x * dist_x + dist_y * dist_y));
 					if (distance >= minrange)
 					{
 						return true;
 					}
 				}
 			}
-			else if (fabs(dist_x) <= maxrange && fabs(dist_y) <= maxrange && dist_both + (dist_both >> 1) <= maxrange)
+			else if (fabs((float)dist_x) <= maxrange && fabs((float)dist_y) <= maxrange && dist_both + (dist_both >> 1) <= maxrange)
 			{
-				distance = (float) sqrt(dist_x * dist_x + dist_y * dist_y);
+				distance = (float) sqrt(float(dist_x * dist_x + dist_y * dist_y));
 				if ((distance <= maxrange) && (distance >= minrange))
 				{
 					return true;
@@ -823,7 +823,7 @@ namespace Game
 			{
 				int nx;
 				dist_x = start_x - x;
-				dist_both = fabs(dist_y) + fabs(dist_y);
+				dist_both = fabs((float)dist_y) + fabs((float)dist_y);
 				for (nx = start_x; nx <= x; nx++, dist_x++, dist_both--)
 				{
 					if (dist_both < range || Distance2D(dist_x, dist_y) <= range)
